research/ai-agent-engine/diagnosis-agent-engine/pyproject.toml
[project]
name = "agent"
version = "0.1.0"

description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <4.0"
dependencies = [
    "numpy>=2.3.2",
    "onnxruntime>=1.22.1",
    "pillow>=11.3.0",
    "requests>=2.32.4",
    "transformers>=4.54.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.5.3",
    "python-dotenv>=1.1.1",
    "aiohttp>=3.12.15",
    "aiofiles>=24.1.0",
    "ipykernel>=6.30.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[dependency-groups]
dev = [
    "langgraph-cli[inmem]>=0.1.71",
    "pytest>=8.3.5",
]
